Added new reducer and action creator
