+++ react/finger-exercises/redux/02 - react-redux/src/redux/book/actions.js
@@ -8,9 +8,6 @@ import { DATA } from '@constants/data';
  */
 export const actions = {
   GET_BOOKS: '@@BOOK/GET_BOOKS',
-  ADD_TO_CART: '@@BOOK/ADD_TO_CART',
-  ADD_ITEM: '@@BOOK/ADD_ITEM',
-  REMOVE_ITEM: '@@BOOK/REMOVE_ITEM',
   SEARCH_ITEM: '@@BOOK/SEARCH_ITEM'
 };
 
@@ -18,18 +15,6 @@ const actionsCreators = {
   getBooks: () => ({
     type: actions.GET_BOOKS,
     payload: DATA
-  }),
-  addToCart: item => ({
-    type: actions.ADD_TO_CART,
-    payload: item
-  }),
-  addItem: itemId => ({
-    type: actions.ADD_ITEM,
-    payload: itemId
-  }),
-  removeItem: itemId => ({
-    type: actions.REMOVE_ITEM,
-    payload: itemId
   }),
   searchBook: value => ({
     type: actions.SEARCH_ITEM,
