+++ react/finger-exercises/redux/02 - react-redux/src/redux/book/reducer.js
@@ -2,7 +2,6 @@ import { actions } from './actions';
 
 const initialState = {
   books: [],
-  bookSelected: [],
   originalData: []
 };
 
@@ -13,28 +12,6 @@ function reducer(state = initialState, action) {
         ...state,
         originalData: [...action.payload],
         books: [...action.payload]
-      };
-    case actions.ADD_TO_CART: // TODO to implement the logic
-      return {
-        ...state,
-        bookSelected: state.bookSelected.concat(action.payload)
-      };
-    case actions.ADD_ITEM: // TODO to implement the logic
-      return {
-        ...state,
-        bookSelected: state.bookSelected.map(elem => {
-          if (elem.id === action.payload) {
-            const aux = { ...elem };
-            aux.quantity += 1;
-            return aux;
-          }
-          return elem;
-        })
-      };
-    case actions.REMOVE_ITEM: // TODO to implement the logic
-      return {
-        ...state,
-        bookSelected: state.bookSelected.filter(actual => actual.id !== action.payload)
       };
     case actions.SEARCH_ITEM: // TODO to implement the logic
       return {
