+++ react/finger-exercises/redux/02 - react-redux/src/redux/cart/actions.js
@@ -1,0 +1,22 @@
+export const actions = {
+  ADD_TO_CART: '@@BOOK/ADD_TO_CART',
+  ADD_ITEM: '@@BOOK/ADD_ITEM',
+  REMOVE_ITEM: '@@BOOK/REMOVE_ITEM'
+};
+
+const actionsCreators = {
+  addToCart: item => ({
+    type: actions.ADD_TO_CART,
+    payload: item
+  }),
+  addItem: itemId => ({
+    type: actions.ADD_ITEM,
+    payload: itemId
+  }),
+  removeItem: itemId => ({
+    type: actions.REMOVE_ITEM,
+    payload: itemId
+  })
+};
+
+export default actionsCreators;
